refactor: extract config update extension to Config.kt
