+++ src/main/kotlin/com/neo/envmanager/command/Delete.kt
@@ -7,13 +7,16 @@ import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.mordant.terminal.YesNoPrompt
+import com.neo.envmanager.com.neo.envmanager.util.extension.update
 import com.neo.envmanager.core.Command
 import com.neo.envmanager.exception.Cancel
 import com.neo.envmanager.exception.error.EnvironmentNotFound
 import com.neo.envmanager.exception.error.SpecifyEnvironmentError
 import com.neo.envmanager.model.Config
 import com.neo.envmanager.model.Environment
-import com.neo.envmanager.util.extension.*
+import com.neo.envmanager.util.extension.deleteChildren
+import com.neo.envmanager.util.extension.requireInstall
+import com.neo.envmanager.util.extension.success
 
 class Delete : Command(
     help = "Delete one or more environments"
