+++ src/main/kotlin/com/neo/envmanager/command/Rename.kt
@@ -1,10 +1,10 @@
 package com.neo.envmanager.command
 
 import com.github.ajalt.clikt.parameters.arguments.argument
+import com.neo.envmanager.com.neo.envmanager.util.extension.update
 import com.neo.envmanager.core.Command
 import com.neo.envmanager.model.Environment
 import com.neo.envmanager.util.extension.requireInstall
-import com.neo.envmanager.util.extension.update
 
 class Rename : Command(
     help = "Rename an environment"
@@ -22,7 +22,7 @@ class Rename : Command(
 
         val config = requireInstall()
 
-         Environment.get(paths.environmentsDir, oldTag).renameTo(newTag)
+        Environment.get(paths.environmentsDir, oldTag).renameTo(newTag)
 
         if (oldTag == config.currentEnv) {
             config.update {
